+++ processing/src/main/java/io/druid/query/aggregation/first/LongFirstBufferAggregator.java
@@ -1,0 +1,81 @@
+/*
+ * Licensed to Metamarkets Group Inc. (Metamarkets) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. Metamarkets licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.druid.query.aggregation.first;
+
+import com.google.common.primitives.Longs;
+import io.druid.collections.SerializablePair;
+import io.druid.query.aggregation.BufferAggregator;
+import io.druid.segment.LongColumnSelector;
+
+import java.nio.ByteBuffer;
+
+public class LongFirstBufferAggregator implements BufferAggregator
+{
+  private final LongColumnSelector timeSelector;
+  private final LongColumnSelector valueSelector;
+
+  public LongFirstBufferAggregator(LongColumnSelector timeSelector, LongColumnSelector valueSelector)
+  {
+    this.timeSelector = timeSelector;
+    this.valueSelector = valueSelector;
+  }
+
+  @Override
+  public void init(ByteBuffer buf, int position)
+  {
+    buf.putLong(position, Long.MAX_VALUE);
+    buf.putLong(position + Longs.BYTES, 0);
+  }
+
+  @Override
+  public void aggregate(ByteBuffer buf, int position)
+  {
+    long time = timeSelector.get();
+    long firstTime = buf.getLong(position);
+    if (time < firstTime) {
+      buf.putLong(position, time);
+      buf.putLong(position + Longs.BYTES, valueSelector.get());
+    }
+  }
+
+  @Override
+  public Object get(ByteBuffer buf, int position)
+  {
+    return new SerializablePair<>(buf.getLong(position), buf.getLong(position + Longs.BYTES));
+  }
+
+  @Override
+  public float getFloat(ByteBuffer buf, int position)
+  {
+    return (float) buf.getLong(position + Longs.BYTES);
+  }
+
+  @Override
+  public long getLong(ByteBuffer buf, int position)
+  {
+    return buf.getLong(position + Longs.BYTES);
+  }
+
+  @Override
+  public void close()
+  {
+    // no resources to cleanup
+  }
+}
