+++ processing/src/main/java/io/druid/query/aggregation/last/LongLastBufferAggregator.java
@@ -17,10 +17,11 @@
  * under the License.
  */
 
-package io.druid.query.aggregation;
+package io.druid.query.aggregation.last;
 
 import com.google.common.primitives.Longs;
 import io.druid.collections.SerializablePair;
+import io.druid.query.aggregation.BufferAggregator;
 import io.druid.segment.LongColumnSelector;
 
 import java.nio.ByteBuffer;
