+++ processing/src/test/java/io/druid/query/aggregation/FirstAggregationTest.java
@@ -22,6 +22,11 @@ package io.druid.query.aggregation;
 import com.metamx.common.Pair;
 import io.druid.collections.SerializablePair;
 import io.druid.jackson.DefaultObjectMapper;
+import io.druid.query.aggregation.first.DoubleFirstAggregator;
+import io.druid.query.aggregation.first.DoubleFirstBufferAggregator;
+import io.druid.query.aggregation.first.FirstAggregatorFactory;
+import io.druid.query.aggregation.first.LongFirstAggregator;
+import io.druid.query.aggregation.first.LongFirstBufferAggregator;
 import io.druid.segment.ColumnSelectorFactory;
 import io.druid.segment.column.Column;
 import org.easymock.EasyMock;
