+++ processing/src/test/java/io/druid/query/aggregation/LastAggregationTest.java
@@ -22,6 +22,11 @@ package io.druid.query.aggregation;
 import com.metamx.common.Pair;
 import io.druid.collections.SerializablePair;
 import io.druid.jackson.DefaultObjectMapper;
+import io.druid.query.aggregation.last.DoubleLastAggregator;
+import io.druid.query.aggregation.last.DoubleLastBufferAggregator;
+import io.druid.query.aggregation.last.LastAggregatorFactory;
+import io.druid.query.aggregation.last.LongLastAggregator;
+import io.druid.query.aggregation.last.LongLastBufferAggregator;
 import io.druid.segment.ColumnSelectorFactory;
 import io.druid.segment.column.Column;
 import org.easymock.EasyMock;
