+++ processing/src/test/java/io/druid/query/timeseries/TimeseriesQueryRunnerTest.java
@@ -36,8 +36,8 @@ import io.druid.query.aggregation.CountAggregatorFactory;
 import io.druid.query.aggregation.DoubleMaxAggregatorFactory;
 import io.druid.query.aggregation.DoubleMinAggregatorFactory;
 import io.druid.query.aggregation.FilteredAggregatorFactory;
-import io.druid.query.aggregation.FirstAggregatorFactory;
-import io.druid.query.aggregation.LastAggregatorFactory;
+import io.druid.query.aggregation.first.FirstAggregatorFactory;
+import io.druid.query.aggregation.last.LastAggregatorFactory;
 import io.druid.query.aggregation.LongSumAggregatorFactory;
 import io.druid.query.aggregation.PostAggregator;
 import io.druid.query.extraction.MapLookupExtractor;
