+++ processing/src/test/java/io/druid/query/topn/TopNQueryRunnerTest.java
@@ -46,8 +46,8 @@ import io.druid.query.aggregation.AggregatorFactory;
 import io.druid.query.aggregation.DoubleMaxAggregatorFactory;
 import io.druid.query.aggregation.DoubleMinAggregatorFactory;
 import io.druid.query.aggregation.FilteredAggregatorFactory;
-import io.druid.query.aggregation.FirstAggregatorFactory;
-import io.druid.query.aggregation.LastAggregatorFactory;
+import io.druid.query.aggregation.first.FirstAggregatorFactory;
+import io.druid.query.aggregation.last.LastAggregatorFactory;
 import io.druid.query.aggregation.PostAggregator;
 import io.druid.query.aggregation.cardinality.CardinalityAggregatorFactory;
 import io.druid.query.aggregation.hyperloglog.HyperUniqueFinalizingPostAggregator;
